refactor: useLocalStorage 변수 useRef화
